[Designer] Fixed issue with ElementGroupIds
... ElementGroup with Id "0" or "-1" would cause a crash due to the way SubElements works. The FileSystem will now ignore elements that have a 0 or -1 as Id

~ "Charge" property in WorldElement changed to use integer ... no reason to have floating point numbers for charges

+++ Assets/ElementDesigner/World/Elements/WorldElement.cs
@@ -20,7 +20,7 @@ public class WorldElement : MonoBehaviour
         }
     }
     // TODO: maybe 'charge' could be determined by weight?
-    public float Charge { get; private set; }
+    public int Charge { get; private set; }
     public Color Color { get; private set; }
     public float MassMultiplier { get; private set; }
